Move TestAnsibleREX CLI tests to Ansible module (#14577)
